% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carga_paquetes.R
\name{carga_paquetes}
\alias{carga_paquetes}
\title{Carga todos los paqutes necesarios}
\usage{
carga_paquetes()
}
\description{
Esta función carga todos los paquetes que se utilizan con relativa frecuencia
}
