% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Familia.R
\docType{data}
\name{Familia}
\alias{Familia}
\title{Number of People in a Family}
\format{
A data frame with 200 observations composed of the
following 3 variables:
\describe{
\item{Id}{Id of the surveyed families.}
\item{Persona}{Number of people in the family.}
\item{Tipo}{Type of Family according to the number of people
that make it up (Normal Family, Average Family, Large Family).}
}
}
\description{
Data set corresponding to the surveys of the number of people
who live in 200 families in Ecuador. Especially in poor areas
of the country.
}
\examples{
library(PaqueteR)
data(Familia)
summary(Familia)
attach(Familia)
str(Familia)

}
\keyword{datasets}
