% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmode.R
\name{getmode}
\alias{getmode}
\title{Title Encontrar la moda en un vector de una serie}
\usage{
getmode(serievector)
}
\arguments{
\item{serievector}{Debes meter un vector alfanumérico}
}
\value{
Me devuelve un valor númerico con la moda
}
\description{
La dunción utiliza aritmetica avazada para sacar la moda de un vector numérico
}
\examples{
getmode(c(1,2,2,2,3))

}
\author{
Guadalupe Ortiz
}
