% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxminR.R
\name{maxminR}
\alias{maxminR}
\title{Maximos y Minimos}
\usage{
maxminR(x)
}
\arguments{
\item{x}{Debes meter una lista con valores numericos}
}
\description{
La función realiza una seleccion del número maximo y minimo a partir de una serie de datos
}
\examples{
maxminR(c(10, 20, 5, 15, 8, 30, -5)

}
\author{
Guadalupe
}
