% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Media.R
\name{Media}
\alias{Media}
\title{Mean of a vector}
\usage{
Media(vector)
}
\arguments{
\item{vector}{A numeric vector.}
}
\description{
Create a function to get the arithmetic mean of a data set. This
object is used to obtain the mean of a vector or numeric
data set.
}
\examples{
library(PaqueteR)
data(Familia)
Media(Familia$Persona)
}
