% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult.R
\name{mult}
\alias{mult}
\title{Multiplicacion entre dos numeros}
\usage{
mult(a, b)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}
}
\value{
Devuelve la multiplicacion entre \code{a} y \code{b}
}
\description{
Esta funcion realiza la multiplicacion entre dos numeros
}
\examples{
mult(5, 2)
}
\seealso{
\code{\link{div}}
}
