% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_NH0437.R
\docType{data}
\name{NH0437}
\alias{NH0437}
\title{Datos meteorológicos de la estación NH0437}
\format{
Un data frame con las columnas:
\describe{
\item{id}{Identificador de la estación (carácter).}
\item{fecha}{Fecha de la medición (Date).}
\item{temperatura_abrigo_150cm}{Temperatura registrada a 150 cm, en °C (numérico).}
}
}
\source{
\url{https://github.com/rse-r/intro-programacion}
}
\usage{
NH0437
}
\description{
Dataset con observaciones diarias de temperatura (150 cm de abrigo) y fecha
correspondientes a la estación NH0437.
}
\keyword{datasets}
