% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Resta entre dos numeros}
\usage{
resta(a, b)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}
}
\value{
Devuelve la resta entre \code{a} y \code{b}
}
\description{
Esta funcion realiza la resta entre dos numeros
}
\examples{
resta(5, 2)
}
\seealso{
\code{\link{suma}}
}
