% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtfun1.R
\name{sqrtfun1}
\alias{sqrtfun1}
\title{Raíz cuadrada 1}
\usage{
sqrtfun1(x)
}
\arguments{
\item{x}{Un vector, o array, numæ”¼ã¸¹rico o complejo}
}
\value{
Raíz cuadrada de x
}
\description{
Función para calcular la raíz cuadrada.
}
\examples{
sqrt(10)
}
