% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtfun2.R
\name{sqrtfun2}
\alias{sqrtfun2}
\title{Raíz cuadrada 2}
\usage{
sqrtfun2(x)
}
\arguments{
\item{x}{Un vector, o array, numæ”¼ã¸¹rico o complejo}
}
\value{
Raíz cuadrada de x
}
\description{
Función para calcular la raíz cuadrada.
}
\examples{
sqrt(10)
}
