% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Suma entre dos numeros}
\usage{
suma(a, b)
}
\arguments{
\item{a}{primer numero}

\item{b}{segundo numero}
}
\value{
Devuelve la suma entre \code{a} y \code{b}
}
\description{
Esta funcion realiza la suma entre dos numeros
}
\examples{
suma(5, 2)
}
\seealso{
\code{\link{resta}}
}
