% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sumatoria.R
\name{Sumatoria}
\alias{Sumatoria}
\title{Mean of a vector}
\usage{
Sumatoria(vector)
}
\arguments{
\item{vector}{A numeric vector.}
}
\description{
Create a function to get the sum of a data set. This object is
used to obtain the sum of a vector or numeric data set.
}
\examples{
library(PaqueteR)
data(Familia)
Sumatoria(Familia$Persona)
}
