% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{test.regresion}
\alias{test.regresion}
\title{Realiza resúmen de un modelo de regresión lineal multiple}
\usage{
test.regresion(x)
}
\arguments{
\item{x}{modelo de regresion}
}
\value{
outputs con test de hipotesis
}
\description{
realiza los test
}
