% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reg.R
\name{TestRegresion}
\alias{TestRegresion}
\title{Realiza resúmen de un modelo de regresión lineal multiple}
\usage{
TestRegresion(x)
}
\arguments{
\item{x}{modelo de regresion}
}
\value{
outputs con test de hipotesis
}
\description{
realiza los test
}
\examples{
u=rnorm(1000,100,12)
normalHist(u);
normalHist(u,dens=TRUE)
normalHist(u,dens=TRUE,col="lightcyan")
}
