% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/div.R
\name{div}
\alias{div}
\title{Division entre dos numeros}
\usage{
div(a, b)
}
\arguments{
\item{a}{numerador}

\item{b}{denominador distinto de cero}
}
\value{
Devuelve la division entre \code{a} y \code{b}
}
\description{
Esta funcion realiza la division entre dos numeros
}
\note{
Cuando b vale 0 la division dara error
}
\examples{
div(5, 2)
}
\seealso{
\code{\link{mult}}
}
