% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Suma.R
\name{Suma}
\alias{Suma}
\title{Add 2 Numbers}
\usage{
Suma(a, b)
}
\arguments{
\item{a}{First number.}

\item{b}{Second number.}
}
\description{
Create a function to calculate the sum of 2 numerical values.
This object is used to add whole numbers or decimals.
}
\examples{
library(PaqueteR)
a <- 10
b <- 2.5
Suma(a,b)
}
